Mod 8.1
index.php is no longer in the root of the project! It has been moved inside the public folder,
for better security and separation of components.
This means that you should configure your web server to "point" to your project's public folder,
and not to the project root. A better practice would be to configure a virtual host to point there.